==> 0_agencies_etl.ipynb <==
# get agencies
import pandas as pd
import requests

BASE_URL = 'https://www.ecfr.gov'
AGENCIES_PATH = '/api/admin/v1/agencies.json'

def get_agencies():
    url = BASE_URL + AGENCIES_PATH
    response = requests.get(url)
    if response.status_code == 200:
        agencies = response.json()
        return agencies
    else:
        print('Error getting agencies')
        return None

agencies = get_agencies()
df_agencies = pd.DataFrame(agencies['agencies'])
df_agencies

# explode out rows for each cfr_reference, add keys of cfr_references obj as col. example value: [{'title': 1, 'chapter': 'III'}]
df_agencies_exploded = df_agencies.explode('cfr_references')
df_agencies_exploded = pd.concat([df_agencies_exploded, df_agencies_exploded['cfr_references'].apply(pd.Series)], axis=1)
df_agencies_exploded.drop(columns=['cfr_references'], inplace=True)
df_agencies_exploded.to_csv('data/ecfr_agencies.csv', index=False)
df_agencies_exploded
