==> 0_corrections_etl.ipynb <==
# get corrections
import pandas as pd
import requests

BASE_URL = 'https://www.ecfr.gov'
AGENCIES_PATH = '/api/admin/v1/corrections.json'

def get_corrections():
    url = BASE_URL + AGENCIES_PATH
    response = requests.get(url)
    response.raise_for_status()
    return response.json()

corrections = get_corrections()
df_corrections = pd.DataFrame(corrections['ecfr_corrections'])
df_corrections

# explode out cfr_references and add keys of cfr_references obj as col names
df_corrections_exploded = df_corrections.explode('cfr_references')
df_corrections_exploded = pd.concat([df_corrections_exploded, df_corrections_exploded['cfr_references'].apply(pd.Series)], axis=1)
df_corrections_exploded = df_corrections_exploded.drop(columns=['cfr_references'])
df_corrections_exploded = pd.concat([df_corrections_exploded, df_corrections_exploded['hierarchy'].apply(pd.Series)], axis=1)
df_corrections_exploded.to_csv('data/ecfr_corrections.csv', index=False)
df_corrections_exploded
