==> 2_word_count_by_agency.ipynb <==
import polars as pl

df_agencies = pl.read_csv("data/ecfr_agencies.csv")
df_word_count_by_chapter = pl.read_csv("data/word_count_by_chapter.csv")

# rename df_agencies title to volume
df_agencies = df_agencies.with_columns([
    pl.col("title").alias("volume")
])

# replace all blank vals with "-"
df_agencies = df_agencies.with_columns([
    pl.col("subchapter").fill_null("-").alias("subchapter"),
    pl.col("part").fill_null("-").alias("part"),
])


print(f'cols for df_agencies: {df_agencies.columns}')
print(f'cols for df_word_count_by_chapter: {df_word_count_by_chapter.columns}')
print(f'column intersection: {set(df_agencies.columns) & set(df_word_count_by_chapter.columns)}')


df_joined = df_agencies.join(df_word_count_by_chapter, on=['volume', 'chapter'], how='inner')
df_joined.write_csv("data/word_count_agency_by_chapter.csv")
df_joined
