==> 3_upload.ipynb <==
import pandas as pd
import sqlalchemy

engine = sqlalchemy.create_engine("postgresql://superset_user:[REDACTED]@doge-db-3.cluster-colq0a2mmskm.us-east-1.rds.amazonaws.com:5432/postgres")

# upload data/ecfr_corrections.csv to the database
df = pd.read_csv("data/ecfr_corrections.csv")

# convert YYYY-MM-DD error_corrected, error_occurred, and last_modified to datetime
df["error_corrected"] = pd.to_datetime(df["error_corrected"])
df["error_occurred"] = pd.to_datetime(df["error_occurred"])
df["last_modified"] = pd.to_datetime(df["last_modified"])

df.to_sql("ecfr_corrections", engine, if_exists="replace", index=False)

# upload data/word_count_agency_by_chapter.csv to the database
df = pd.read_csv("data/word_count_agency_by_chapter.csv")


df.to_sql("word_count_agency_by_chapter", engine, if_exists="replace", index=False)

# upload data/ecfr_contents.csv to the database
df = pd.read_csv("data/ecfr_sections.csv")

# drop descendant_range, parent_id
df.drop(columns=["descendant_range", "parent_id"], inplace=True)

# convert received on (format 2023-04-12T10:20:08-0400) to datetime
df["received_on"] = pd.to_datetime(df["received_on"], format="%Y-%m-%dT%H:%M:%S%z")

df.to_sql("ecfr_sections", engine, if_exists="replace", index=False)

# upload data/word_count_by_volume.csv to the database
df = pd.read_csv("data/word_count_by_volume.csv")

df.to_sql("word_count_by_volume", engine, if_exists="replace", index=False)
